refactor: begin work on upgrading to Python 3
