+++ src/newpantheon/experiments/data/pantheon_metadata.json
@@ -1,1 +1,1 @@
-{"command": "experiment", "experiment_command": "test", "mode": "local", "flows": 1, "runtime": 30, "interval": 0, "run_times": 1, "start_run_id": 1, "random_order": false, "uplink_trace": "12mbps.trace", "downlink_trace": "12mbps.trace", "cc_schemes": "['cubic']", "git_summary": "branch: shinwoo/rewrite-experiment @ 4e1703a191bef110b40edff59df106b1b3f4f6b5\nthird_party/pantheon-tunnel @ f866d3f58d27afd942717625ee3a354cc2e802bd\n"}
+{"command": "experiment", "experiment_command": "test", "mode": "local", "flows": 1, "runtime": 30, "interval": 0, "run_times": 1, "start_run_id": 1, "random_order": false, "uplink_trace": "12mbps.trace", "downlink_trace": "12mbps.trace", "cc_schemes": "['cubic']", "git_summary": "branch: shinwoo/rewrite-experiment @ 4c047ef56cba603684da63e96c71a01727ae1ff7\nthird_party/pantheon-tunnel @ f866d3f58d27afd942717625ee3a354cc2e802bd\n"}
